requirements.txt
fastapi
uvicorn
sqlmodel
ortools
python-dotenv
httpx
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
python-multipart